Add new function, add more functionality, fix bug
First, created a doOperation function to return the correct value of
the operation, because it didn't show the Divided by 0 error before.

Secondly, the use of the (.) works properly now. It only accepts the
entry of one (.), so if the number has a decimal, it doesn't allow the
user to input more.

Finally, fixed a bug where if I pressed the (=) symbol after a had done
an operation, it didn't display the same number.